Add ModifyDoctorView UI for doctor information editing
Created a new `ModifyDoctorView.xaml` file that defines a user interface for modifying doctor information. The layout includes a title, a `DataGrid` for displaying and editing doctor details, a save button, and an error message display. The `DataGrid` is configured to bind to a collection of doctors with relevant attributes such as Doctor ID, User ID, Department ID, Experience, and License Number.