Fixed coins logic & updated part of some minor fixes (#23)
* Updated the dailyloginbonus to give 100 coins

* Removed 2 unused folders from the .csproj

* Made it correctly save the money of the user

* added icons for the app

* Update logo assets in Package.appxmanifest

Replaced existing logo files with a new asset (`icon.jpg`) for the application. Added a lock screen badge logo using the same asset.

* added icon for the tab

* Added user img

* Updated profile settings page to remove the old profile picture

* Updated main window's name

* Update Duo.csproj for resource management and assets

Removed PRIResource entries from the Repositories directory.
Added userIcon.png to the content files with PreserveNewest
copy option to ensure it is included in the output directory.