Fix logout crashes and sing up bugs (#46)
* Logout navigation refactor in MainPage.xaml.cs

Refactored the `LogoutButton_Click` method to utilize a `Frame` for navigation to the `LoginPage` instead of directly setting the content of `MainAppWindow`. This change improves navigation management within the application.

* Refactor user service handling in SignUpPage

Replaced direct access to App.userService with dependency
injection of IUserService for setting the username of the
newly created user. Adjusted the order of operations to
display the success dialog after setting the user,
enhancing user feedback flow.